Release v1.2.2
This is the first release of `stringcase-new`; version numbers are
intended to follow [semantic versioning](https://semver.org) in general,
but are starting here for continuity with the original project. (Version
1.2.1 has been skipped - recognizing that the original author appears to
have intended to make that release but never did.)

This version is intended to be bug compatible with the last published
version (1.2.0) of the original. The only change to the main source code
is to reorganize it; the only change to the test code is to remove a
`sys.path` hack (unneeded, and became incorrect with the
reorganization).

Added an initial CHANGELOG.md to ensure the package metadata points
somewhere valid.